+++ web/app/js/components/Navigation.jsx
@@ -1,37 +1,45 @@
-import { githubIcon, linkerdWordLogo, slackIcon } from './util/SvgWrappers.jsx';
+import { daemonsetIcon, deploymentIcon, githubIcon, jobIcon, linkerdWordLogo, namespaceIcon, podIcon, replicaSetIcon, slackIcon, statefulSetIcon } from './util/SvgWrappers.jsx';
 import AppBar from '@material-ui/core/AppBar';
+import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown';
 import Badge from '@material-ui/core/Badge';
 import BreadcrumbHeader from './BreadcrumbHeader.jsx';
+import Button from '@material-ui/core/Button';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import CloudQueueIcon from '@material-ui/icons/CloudQueue';
+import Dialog from '@material-ui/core/Dialog';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogContentText from '@material-ui/core/DialogContentText';
+import DialogTitle from '@material-ui/core/DialogTitle';
 import Divider from '@material-ui/core/Divider';
 import Drawer from '@material-ui/core/Drawer';
 import EmailIcon from '@material-ui/icons/Email';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import HomeIcon from '@material-ui/icons/Home';
 import IconButton from '@material-ui/core/IconButton';
 import LibraryBooksIcon from '@material-ui/icons/LibraryBooks';
 import { Link } from 'react-router-dom';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
-import MenuIcon from '@material-ui/icons/Menu';
+import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import MenuList from '@material-ui/core/MenuList';
-import NavigationResources from './NavigationResources.jsx';
 import PropTypes from 'prop-types';
 import React from 'react';
 import ReactRouterPropTypes from 'react-router-prop-types';
-import SentimentVerySatisfiedIcon from '@material-ui/icons/SentimentVerySatisfied';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Version from './Version.jsx';
 import _maxBy from 'lodash/maxBy';
 import classNames from 'classnames';
+import { faBars } from '@fortawesome/free-solid-svg-icons/faBars';
+import { faCloud } from '@fortawesome/free-solid-svg-icons/faCloud';
 import { faExternalLinkAlt } from '@fortawesome/free-solid-svg-icons/faExternalLinkAlt';
+import { faFilter } from '@fortawesome/free-solid-svg-icons/faFilter';
 import { faMicroscope } from '@fortawesome/free-solid-svg-icons/faMicroscope';
 import { faRandom } from '@fortawesome/free-solid-svg-icons/faRandom';
+import { faSmile } from '@fortawesome/free-regular-svg-icons/faSmile';
 import { faStream } from '@fortawesome/free-solid-svg-icons/faStream';
 import grey from '@material-ui/core/colors/grey';
+import { processSingleResourceRollup } from './util/MetricUtils.jsx';
 import { withContext } from './util/AppContext.jsx';
 import { withStyles } from '@material-ui/core/styles';
 import yellow from '@material-ui/core/colors/yellow';
@@ -41,7 +49,7 @@ const localStorageKey = "linkerd-updates-last-clicked";
 const minBrowserWidth = 960;
 
 const styles = theme => {
-  const drawerWidth = theme.spacing.unit * 31;
+  const drawerWidth = theme.spacing.unit * 36;
   const drawerWidthClosed = theme.spacing.unit * 9;
   const navLogoWidth = theme.spacing.unit * 22.5;
   const contentPadding = theme.spacing.unit * 3;
@@ -64,7 +72,6 @@ const styles = theme => {
     },
     appBar: {
       position: "absolute",
-      width: `calc(100% - ${drawerWidthClosed}px)`,
       color: 'white',
       transition: leaving,
     },
@@ -122,18 +129,37 @@ const styles = theme => {
       marginLeft: `-${navLogoWidth+theme.spacing.unit/2}px`,
       transition: leavingFn(['margin', 'opacity']),
     },
+    namespaceChangeButton: {
+      marginLeft: `${drawerWidth * .075}px`,
+      marginTop: "11px",
+      width: `${drawerWidth * .8}px`,
+    },
     navMenuItem: {
       paddingLeft: `${contentPadding}px`,
       paddingRight: `${contentPadding}px`,
     },
     shrinkIcon: {
-      fontSize: "18px",
+      fontSize: "19px",
       paddingLeft: "3px",
       paddingRight: "3px",
+    },
+    shrinkCloudIcon: {
+      fontSize: "18px",
+      paddingLeft: "1px",
     },
     // color is consistent with Octopus Graph coloring
     externalLinkIcon: {
       color: grey[500],
+    },
+    sidebarHeading: {
+      color: grey[500],
+      outline: "none",
+      paddingTop: "9px",
+      paddingBottom: "9px",
+      marginLeft: `${drawerWidth * .09}px`,
+    },
+    toggleDrawerButton: {
+      marginRight: `${contentPadding}px`,
     },
     badge: {
       backgroundColor: yellow[500],
@@ -146,32 +172,77 @@ class NavigationBase extends React.Component {
     super(props);
     this.api = this.props.api;
     this.handleApiError = this.handleApiError.bind(this);
+    this.handleConfirmNamespaceChange = this.handleConfirmNamespaceChange.bind(this);
     this.handleCommunityClick = this.handleCommunityClick.bind(this);
+    this.handleDialogCancel = this.handleDialogCancel.bind(this);
+    this.handleNamespaceMenuClick = this.handleNamespaceMenuClick.bind(this);
     this.updateWindowDimensions = this.updateWindowDimensions.bind(this);
 
     this.state = this.getInitialState();
+    this.loadFromServer = this.loadFromServer.bind(this);
   }
 
   getInitialState() {
     return {
+      anchorEl: null,
       drawerOpen: true,
-      helpMenuOpen: false,
+      namespaceMenuOpen: false,
       hideUpdateBadge: true,
       latestVersion: '',
       isLatest: true,
-      namespaceFilter: "all"
+      namespaces: [],
+      pendingRequests: false,
+      pollingInterval: 10000,
+      loaded: false,
+      error: null,
+      showNamespaceChangeDialog: false,
     };
   }
 
   componentDidMount() {
+    this.loadFromServer();
+    this.timerId = window.setInterval(this.loadFromServer, this.state.pollingInterval);
     this.fetchVersion();
     this.fetchLatestCommunityUpdate();
     this.updateWindowDimensions();
     window.addEventListener("resize", this.updateWindowDimensions);
   }
 
+  componentWillUpdate() {
+    if (this.props.history) {
+      this.props.checkNamespaceMatch(this.props.history.location.pathname);
+    }
+  }
+
   componentWillUnmount() {
     window.removeEventListener("resize", this.updateWindowDimensions);
+    window.clearInterval(this.timerId);
+    this.api.cancelCurrentRequests();
+  }
+
+  // API returns namespaces for namespace select button. No metrics returned.
+  loadFromServer() {
+    if (this.state.pendingRequests) {
+      return;
+    }
+    this.setState({ pendingRequests: true });
+
+    let apiRequests = [
+      this.api.fetchMetrics(this.api.urlsForResourceNoStats("namespace"))
+    ];
+
+    this.api.setCurrentRequests(apiRequests);
+
+    Promise.all(this.api.getCurrentPromises())
+      .then(([allNs]) => {
+        let namespaces = processSingleResourceRollup(allNs);
+        this.setState({
+          namespaces,
+          pendingRequests: false,
+          error: null
+        });
+      })
+      .catch(this.handleApiError);
   }
 
   fetchVersion() {
@@ -224,12 +295,46 @@ class NavigationBase extends React.Component {
     this.setState({ hideUpdateBadge: true });
   }
 
+  handleDialogCancel = () => {
+    this.setState({showNamespaceChangeDialog: false});
+  }
+
   handleDrawerClick = () => {
     this.setState(state => ({ drawerOpen: !state.drawerOpen }));
   };
 
-  handleHelpMenuClick = () => {
-    this.setState(state => ({ helpMenuOpen: !state.helpMenuOpen }));
+  handleConfirmNamespaceChange = () => {
+    this.setState({showNamespaceChangeDialog: false});
+    this.props.updateNamespaceInContext(this.state.newNamespace);
+    this.props.history.push(`/namespaces/${this.state.newNamespace}`);
+  }
+
+  handleNamespaceChange = namespace => {
+    this.setState({ namespaceMenuOpen: false });
+    if (namespace === this.props.selectedNamespace) {
+      return;
+    }
+    let path = this.props.history.location.pathname;
+    let pathParts = path.split("/");
+    if (pathParts.length === 3 || pathParts.length === 4) {
+      // path is /namespaces/someNamespace/resourceType
+      //      or /namespaces/someNamespace
+      path = path.replace(this.props.selectedNamespace, namespace);
+      this.props.history.push(path);
+      this.props.updateNamespaceInContext(namespace);
+    } else if (pathParts.length === 5) {
+      // path is /namespace/someNamespace/resourceType/someResource
+      this.setState({ showNamespaceChangeDialog: true,
+        newNamespace: namespace });
+    } else {
+      // update the selectedNamespace in context with no path changes
+      this.props.updateNamespaceInContext(namespace);
+    }
+  }
+
+  handleNamespaceMenuClick = event => {
+    this.setState({ anchorEl: event.currentTarget });
+    this.setState(state => ({ namespaceMenuOpen: !state.namespaceMenuOpen }));
   }
 
   menuItem(path, title, icon, onClick) {
@@ -260,62 +365,174 @@ class NavigationBase extends React.Component {
   }
 
   render() {
-    const { classes, ChildComponent, ...otherProps } = this.props;
+    const { api, classes, selectedNamespace, ChildComponent, ...otherProps } = this.props;
+    const { namespaces, anchorEl, drawerOpen, showNamespaceChangeDialog, newNamespace } = this.state;
+    let formattedNamespaceName = selectedNamespace;
+    if (formattedNamespaceName === "_all") {
+      formattedNamespaceName = "All Namespaces";
+    }
 
     return (
       <div className={classes.root}>
         <AppBar
-          className={classNames(classes.appBar, {[classes.appBarShift]: this.state.drawerOpen} )}>
+          className={classNames(classes.appBar, {[classes.appBarShift]: drawerOpen} )}>
           <Toolbar>
             <Typography variant="h6" color="inherit" noWrap>
+              { !drawerOpen &&
+              <IconButton className={classes.toggleDrawerButton}>
+                <FontAwesomeIcon icon={faBars} onClick={this.handleDrawerClick} />
+              </IconButton>
+              }
               <BreadcrumbHeader {...this.props} />
             </Typography>
           </Toolbar>
         </AppBar>
 
         <Drawer
-          className={classNames(classes.drawer, {[classes.drawerClose]: !this.state.drawerOpen} )}
-          variant="permanent"
+          className={classNames(classes.drawer, {[classes.drawerClose]: !drawerOpen} )}
+          variant="persistent"
           classes={{
-            paper: classNames(classes.drawerPaper, {[classes.drawerPaperClose]: !this.state.drawerOpen} ),
+            paper: classNames(classes.drawerPaper, {[classes.drawerPaperClose]: !drawerOpen} ),
           }}
-          open={this.state.drawerOpen}>
+          open={drawerOpen}>
           <div className={classNames(classes.navToolbar)}>
-            <div className={classNames(classes.linkerdNavLogo, {[classes.linkerdNavLogoClose]: !this.state.drawerOpen} )}>
-              <Link to="/overview">{linkerdWordLogo}</Link>
+            <div className={classNames(classes.linkerdNavLogo, {[classes.linkerdNavLogoClose]: !drawerOpen} )}>
+              <Link to="/namespaces">{linkerdWordLogo}</Link>
             </div>
             <IconButton className="drawer-toggle-btn" onClick={this.handleDrawerClick}>
-              {this.state.drawerOpen ? <ChevronLeftIcon /> : <MenuIcon />}
+              <ChevronLeftIcon />
             </IconButton>
           </div>
 
           <Divider />
-
           <MenuList>
-            { this.menuItem("/overview", "Overview", <HomeIcon />) }
-            { this.menuItem("/tap", "Tap", <FontAwesomeIcon icon={faMicroscope} className={classes.shrinkIcon} />) }
-            { this.menuItem("/top", "Top", <FontAwesomeIcon icon={faStream} className={classes.shrinkIcon} />) }
-            { this.menuItem("/routes", "Top Routes", <FontAwesomeIcon icon={faRandom} className={classes.shrinkIcon} />) }
-            { this.menuItem("/servicemesh", "Service Mesh", <CloudQueueIcon className={classes.shrinkIcon} />) }
-            <NavigationResources />
+            <Typography variant="button" className={classes.sidebarHeading}>
+                Cluster
+            </Typography>
+
+            { this.menuItem("/namespaces", "Namespaces", namespaceIcon) }
+
+
+            { this.menuItem("/controlplane", "Control Plane",
+              <FontAwesomeIcon icon={faCloud} className={classes.shrinkCloudIcon} />) }
+
           </MenuList>
 
           <Divider />
 
           <MenuList>
-            <MenuItem component="a" href="https://linkerd.io/2/overview/" target="_blank" className={classes.navMenuItem}>
-              <ListItemIcon><LibraryBooksIcon className={classes.shrinkIcon} /></ListItemIcon>
-              <ListItemText primary="Documentation" />
-              <FontAwesomeIcon icon={faExternalLinkAlt} className={classes.externalLinkIcon} size="xs" />
-            </MenuItem>
+            <Button
+              variant="contained"
+              className={classes.namespaceChangeButton}
+              size="large"
+              onClick={this.handleNamespaceMenuClick}>
+              { formattedNamespaceName }
+              <ArrowDropDownIcon />
+            </Button>
+            <Menu
+              anchorEl={anchorEl}
+              open={this.state.namespaceMenuOpen}
+              keepMounted
+              onClose={this.handleNamespaceMenuClick}>
+              <MenuItem
+                value="all"
+                onClick={() => this.handleNamespaceChange("_all")}>
+                  All Namespaces
+              </MenuItem>
+
+              <Divider />
+
+              {namespaces.map(ns => (
+                <MenuItem
+                  onClick={() => this.handleNamespaceChange(ns.name)}
+                  key={ns.name}>
+                  {ns.name}
+                </MenuItem>
+              ))}
+            </Menu>
+            <Dialog
+              open={showNamespaceChangeDialog}
+              onClose={this.handleClose}
+              aria-labelledby="form-dialog-title">
+              <DialogTitle id="form-dialog-title">
+                Change namespace?
+              </DialogTitle>
+              <DialogContent>
+                <DialogContentText>
+              The resource you are viewing is in a different namespace than the namespace you have selected. Do you want to change the namespace from { selectedNamespace } to { newNamespace }?
+                </DialogContentText>
+              </DialogContent>
+              <DialogActions>
+                <Button onClick={this.handleConfirmNamespaceChange} color="primary">
+                    Yes
+                </Button>
+                <Button onClick={this.handleDialogCancel} variant="text">
+                  No
+                </Button>
+              </DialogActions>
+            </Dialog>
+          </MenuList>
+
+          <MenuList>
+            <Typography variant="button" className={classes.sidebarHeading}>
+                Workloads
+            </Typography>
+
+            { this.menuItem(`/namespaces/${selectedNamespace}/daemonsets`, "Daemon Sets", daemonsetIcon) }
+
+            { this.menuItem(`/namespaces/${selectedNamespace}/deployments`, "Deployments", deploymentIcon) }
+
+            { this.menuItem(`/namespaces/${selectedNamespace}/jobs`, "Jobs", jobIcon) }
+
+            { this.menuItem(`/namespaces/${selectedNamespace}/pods`, "Pods", podIcon) }
+
+            { this.menuItem(`/namespaces/${selectedNamespace}/replicationcontrollers`, "Replication Controllers", replicaSetIcon) }
+
+            { this.menuItem(`/namespaces/${selectedNamespace}/statefulsets`, "Stateful Sets", statefulSetIcon) }
+          </MenuList>
+
+          <MenuList>
+            <Typography variant="button" className={classes.sidebarHeading}>
+                Configuration
+            </Typography>
+
+            { this.menuItem(`/namespaces/${selectedNamespace}/trafficsplits`, "Traffic Splits", <FontAwesomeIcon icon={faFilter} className={classes.shrinkIcon} />) }
+
+          </MenuList>
+          <Divider />
+          <MenuList >
+            <Typography variant="button" className={classes.sidebarHeading}>
+                Tools
+            </Typography>
+
+            { this.menuItem("/tap", "Tap", <FontAwesomeIcon icon={faMicroscope} className={classes.shrinkIcon} />) }
+            { this.menuItem("/top", "Top", <FontAwesomeIcon icon={faStream} className={classes.shrinkIcon} />) }
+            { this.menuItem("/routes", "Routes", <FontAwesomeIcon icon={faRandom} className={classes.shrinkIcon} />) }
+
+          </MenuList>
+          <Divider />
+          <MenuList>
             { this.menuItem("/community", "Community",
               <Badge
                 classes={{ badge: classes.badge }}
                 invisible={this.state.hideUpdateBadge}
                 badgeContent="1">
-                <SentimentVerySatisfiedIcon className={classes.shrinkIcon} />
+                <FontAwesomeIcon icon={faSmile} className={classes.shrinkIcon} />
               </Badge>, this.handleCommunityClick
               ) }
+
+            <MenuItem component="a" href="https://linkerd.io/2/overview/" target="_blank" className={classes.navMenuItem}>
+              <ListItemIcon><LibraryBooksIcon className={classes.shrinkIcon} /></ListItemIcon>
+              <ListItemText primary="Documentation" />
+              <FontAwesomeIcon icon={faExternalLinkAlt} className={classes.externalLinkIcon} size="xs" />
+            </MenuItem>
+
+            <MenuItem component="a" href="https://github.com/linkerd/linkerd2/issues/new/choose" target="_blank" className={classes.navMenuItem}>
+              <ListItemIcon>{githubIcon}</ListItemIcon>
+              <ListItemText primary="GitHub" />
+              <FontAwesomeIcon icon={faExternalLinkAlt} className={classes.externalLinkIcon} size="xs" />
+            </MenuItem>
+
             <MenuItem component="a" href="https://lists.cncf.io/g/cncf-linkerd-users" target="_blank" className={classes.navMenuItem}>
               <ListItemIcon><EmailIcon className={classes.shrinkIcon} /></ListItemIcon>
               <ListItemText primary="Mailing List" />
@@ -328,15 +545,10 @@ class NavigationBase extends React.Component {
               <FontAwesomeIcon icon={faExternalLinkAlt} className={classes.externalLinkIcon} size="xs" />
             </MenuItem>
 
-            <MenuItem component="a" href="https://github.com/linkerd/linkerd2/issues/new/choose" target="_blank" className={classes.navMenuItem}>
-              <ListItemIcon>{githubIcon}</ListItemIcon>
-              <ListItemText primary="File an Issue" />
-              <FontAwesomeIcon icon={faExternalLinkAlt} className={classes.externalLinkIcon} size="xs" />
-            </MenuItem>
           </MenuList>
 
           {
-            !this.state.drawerOpen ? null : <Version
+            !drawerOpen ? null : <Version
               isLatest={this.state.isLatest}
               latestVersion={this.state.latestVersion}
               releaseVersion={this.props.releaseVersion}
@@ -345,7 +557,7 @@ class NavigationBase extends React.Component {
           }
         </Drawer>
 
-        <main className={classNames(classes.content, {[classes.contentDrawerClose]: !this.state.drawerOpen})}>
+        <main className={classNames(classes.content, {[classes.contentDrawerClose]: !drawerOpen})}>
           <div className={classes.toolbar} />
           <div className="main-content"><ChildComponent {...otherProps} /></div>
         </main>